[WIP] Track exec and javaexec calls during configuration time
External processes are problematic for the configuration cache because
there is no reliable way to figure out if the call affects the
configuration. This CL intercepts all calls to exec() and javaexec()
methods (of Scripts, Project and ExecOperations) during configuration
time and emits configuration cache errors. Non-user code calls to the
process execution infrastructure (e.g. worker spawn or innards of Exec
tasks) are not intercepted.

[WIP NOTE]: Tests are incomplete and should cover more varieties of
the API - init scripts, plugins. The tests should also cover cases
where execution should not be detected.